give sources a scitype and mv code to new files sources.jl
fix
